requirements.txt
# Healthcare AI Demo Requirements
# These packages are available in the Snowflake Anaconda channel
# https://repo.anaconda.com/pkgs/snowflake/

streamlit>=1.28.0
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
altair>=5.0.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
tomli>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# For local development only (not needed in Snowflake)
# watchdog>=3.0.0